+++ src/lib.rs
@@ -8,21 +8,13 @@ extern crate alloc;
 
 /// Traits that allow chaining parser outputs together.
 pub mod chain;
-/// Combinators that allow combining and extending existing parsers.
 pub mod combinator;
-/// Utilities for debugging parsers.
 pub mod debug;
-/// Error types, traits and utilities.
 pub mod error;
-/// Parser primitives that accept specific token patterns.
 pub mod primitive;
-/// Types and traits that facilitate error recovery.
 pub mod recovery;
-/// Recursive parsers (parser that include themselves within their patterns).
 pub mod recursive;
-/// Types and traits related to spans.
 pub mod span;
-/// Token streams and behaviours.
 pub mod stream;
 pub mod text;
 
@@ -64,6 +56,9 @@ use std::{
 };
 
 /// Commonly used functions, traits and types.
+///
+/// *Listen, three eyes,” he said, “don’t you try to outweird me, I get stranger things than you free with my breakfast
+/// cereal.”*
 pub mod prelude {
     pub use super::{
         error::{Error as _, Simple},
@@ -415,6 +410,8 @@ pub trait Parser<I: Clone, O> {
 
     /// Validate an output, producing non-terminal errors if it does not fulfil certain criteria.
     ///
+    /// This function also permits mapping the output to a value of another type, similar to [`Parser::map`].
+    ///
     /// If you wish parsing of this pattern to halt when an error is generated instead of continuing, consider using
     /// [`Parser::try_map`] instead.
     ///
@@ -435,12 +432,12 @@ pub trait Parser<I: Clone, O> {
     /// assert_eq!(large_int.parse("537"), Ok(537));
     /// assert!(large_int.parse("243").is_err());
     /// ```
-    fn validate<F>(self, f: F) -> Validate<Self, F>
+    fn validate<F, U>(self, f: F) -> Validate<Self, O, F>
     where
         Self: Sized,
-        F: Fn(O, <Self::Error as Error<I>>::Span, &mut dyn FnMut(Self::Error)) -> O,
+        F: Fn(O, <Self::Error as Error<I>>::Span, &mut dyn FnMut(Self::Error)) -> U,
     {
-        Validate(self, f)
+        Validate(self, f, PhantomData)
     }
 
     /// Label the pattern parsed by this parser for more useful error messages.